splitted binary and library
