+++ psykoosi_lib/analysis.cpp
