+++ psykoosi_lib/analysis.h
