+++ psykoosi_lib/disassemble.cpp
