+++ psykoosi_lib/disassemble.h
