+++ psykoosi_lib/fileformats.cpp
