+++ psykoosi_lib/loading.cpp
