+++ psykoosi_lib/logging.cpp
