+++ psykoosi_lib/logging.h
