+++ psykoosi_lib/rebuild.cpp
