+++ psykoosi_lib/rebuild.h
