+++ psykoosi_lib/structures.h
