+++ psykoosi_lib/virtualmemory.cpp
