+++ psykoosi_lib/virtualmemory.h
